fix: update bubble size on change of slider
The problem was visibility of the functions/variables (like svg), so moved function calls and definitions into the main chart function. The radius updates should work now. I'm not sure if this is idiomatic D3 though, might be good for extra credit to find an existing example that combines all of: reusable charts (that big chart function), force layout, and updates on slider changes. But make it work before making it pretty, amirite?